+++ lib/src/infrastructure/session_repository.dart
@@ -1,6 +1,3 @@
-import 'dart:convert';
-
-import 'package:ankigpt/src/models/get_cards_response.dart';
 import 'package:ankigpt/src/models/session_id.dart';
 import 'package:ankigpt/src/providers/buy_provider.dart';
 import 'package:cloud_functions/cloud_functions.dart';
@@ -24,19 +21,5 @@ class SessionRepository {
       }
     });
     return result.data['id'];
-  }
-
-  Future<GetCardsResponse> getCards({
-    required SessionId sessionId,
-  }) async {
-    final result = await functions
-        .httpsCallableFromUrl(routeFunctionsUrl)
-        .call<Map<String, dynamic>>({
-      'destination': 'getCards',
-      'payload': {
-        'sessionId': sessionId,
-      }
-    });
-    return GetCardsResponse.fromJson(jsonDecode(jsonEncode(result.data)));
   }
 }
