--- lib/src/models/get_cards_response.freezed.dart
@@ -1,294 +1,0 @@
-// coverage:ignore-file
-// GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: type=lint
-// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
-
-part of 'get_cards_response.dart';
-
-// **************************************************************************
-// FreezedGenerator
-// **************************************************************************
-
-T _$identity<T>(T value) => value;
-
-final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
-
-GetCardsResponse _$GetCardsResponseFromJson(Map<String, dynamic> json) {
-  return _GetCardsResponse.fromJson(json);
-}
-
-/// @nodoc
-mixin _$GetCardsResponse {
-  String? get sessionId => throw _privateConstructorUsedError;
-  List<AnkiCard>? get cards => throw _privateConstructorUsedError;
-  bool get isCompleted => throw _privateConstructorUsedError;
-  CsvMetadata? get csv => throw _privateConstructorUsedError;
-  @JsonKey(fromJson: parseError)
-  String? get error => throw _privateConstructorUsedError;
-  Language? get language => throw _privateConstructorUsedError;
-
-  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
-  @JsonKey(ignore: true)
-  $GetCardsResponseCopyWith<GetCardsResponse> get copyWith =>
-      throw _privateConstructorUsedError;
-}
-
-/// @nodoc
-abstract class $GetCardsResponseCopyWith<$Res> {
-  factory $GetCardsResponseCopyWith(
-          GetCardsResponse value, $Res Function(GetCardsResponse) then) =
-      _$GetCardsResponseCopyWithImpl<$Res, GetCardsResponse>;
-  @useResult
-  $Res call(
-      {String? sessionId,
-      List<AnkiCard>? cards,
-      bool isCompleted,
-      CsvMetadata? csv,
-      @JsonKey(fromJson: parseError) String? error,
-      Language? language});
-
-  $CsvMetadataCopyWith<$Res>? get csv;
-}
-
-/// @nodoc
-class _$GetCardsResponseCopyWithImpl<$Res, $Val extends GetCardsResponse>
-    implements $GetCardsResponseCopyWith<$Res> {
-  _$GetCardsResponseCopyWithImpl(this._value, this._then);
-
-  // ignore: unused_field
-  final $Val _value;
-  // ignore: unused_field
-  final $Res Function($Val) _then;
-
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? sessionId = freezed,
-    Object? cards = freezed,
-    Object? isCompleted = null,
-    Object? csv = freezed,
-    Object? error = freezed,
-    Object? language = freezed,
-  }) {
-    return _then(_value.copyWith(
-      sessionId: freezed == sessionId
-          ? _value.sessionId
-          : sessionId // ignore: cast_nullable_to_non_nullable
-              as String?,
-      cards: freezed == cards
-          ? _value.cards
-          : cards // ignore: cast_nullable_to_non_nullable
-              as List<AnkiCard>?,
-      isCompleted: null == isCompleted
-          ? _value.isCompleted
-          : isCompleted // ignore: cast_nullable_to_non_nullable
-              as bool,
-      csv: freezed == csv
-          ? _value.csv
-          : csv // ignore: cast_nullable_to_non_nullable
-              as CsvMetadata?,
-      error: freezed == error
-          ? _value.error
-          : error // ignore: cast_nullable_to_non_nullable
-              as String?,
-      language: freezed == language
-          ? _value.language
-          : language // ignore: cast_nullable_to_non_nullable
-              as Language?,
-    ) as $Val);
-  }
-
-  @override
-  @pragma('vm:prefer-inline')
-  $CsvMetadataCopyWith<$Res>? get csv {
-    if (_value.csv == null) {
-      return null;
-    }
-
-    return $CsvMetadataCopyWith<$Res>(_value.csv!, (value) {
-      return _then(_value.copyWith(csv: value) as $Val);
-    });
-  }
-}
-
-/// @nodoc
-abstract class _$$_GetCardsResponseCopyWith<$Res>
-    implements $GetCardsResponseCopyWith<$Res> {
-  factory _$$_GetCardsResponseCopyWith(
-          _$_GetCardsResponse value, $Res Function(_$_GetCardsResponse) then) =
-      __$$_GetCardsResponseCopyWithImpl<$Res>;
-  @override
-  @useResult
-  $Res call(
-      {String? sessionId,
-      List<AnkiCard>? cards,
-      bool isCompleted,
-      CsvMetadata? csv,
-      @JsonKey(fromJson: parseError) String? error,
-      Language? language});
-
-  @override
-  $CsvMetadataCopyWith<$Res>? get csv;
-}
-
-/// @nodoc
-class __$$_GetCardsResponseCopyWithImpl<$Res>
-    extends _$GetCardsResponseCopyWithImpl<$Res, _$_GetCardsResponse>
-    implements _$$_GetCardsResponseCopyWith<$Res> {
-  __$$_GetCardsResponseCopyWithImpl(
-      _$_GetCardsResponse _value, $Res Function(_$_GetCardsResponse) _then)
-      : super(_value, _then);
-
-  @pragma('vm:prefer-inline')
-  @override
-  $Res call({
-    Object? sessionId = freezed,
-    Object? cards = freezed,
-    Object? isCompleted = null,
-    Object? csv = freezed,
-    Object? error = freezed,
-    Object? language = freezed,
-  }) {
-    return _then(_$_GetCardsResponse(
-      sessionId: freezed == sessionId
-          ? _value.sessionId
-          : sessionId // ignore: cast_nullable_to_non_nullable
-              as String?,
-      cards: freezed == cards
-          ? _value._cards
-          : cards // ignore: cast_nullable_to_non_nullable
-              as List<AnkiCard>?,
-      isCompleted: null == isCompleted
-          ? _value.isCompleted
-          : isCompleted // ignore: cast_nullable_to_non_nullable
-              as bool,
-      csv: freezed == csv
-          ? _value.csv
-          : csv // ignore: cast_nullable_to_non_nullable
-              as CsvMetadata?,
-      error: freezed == error
-          ? _value.error
-          : error // ignore: cast_nullable_to_non_nullable
-              as String?,
-      language: freezed == language
-          ? _value.language
-          : language // ignore: cast_nullable_to_non_nullable
-              as Language?,
-    ));
-  }
-}
-
-/// @nodoc
-@JsonSerializable()
-class _$_GetCardsResponse implements _GetCardsResponse {
-  const _$_GetCardsResponse(
-      {this.sessionId,
-      final List<AnkiCard>? cards,
-      required this.isCompleted,
-      this.csv,
-      @JsonKey(fromJson: parseError) this.error,
-      this.language})
-      : _cards = cards;
-
-  factory _$_GetCardsResponse.fromJson(Map<String, dynamic> json) =>
-      _$$_GetCardsResponseFromJson(json);
-
-  @override
-  final String? sessionId;
-  final List<AnkiCard>? _cards;
-  @override
-  List<AnkiCard>? get cards {
-    final value = _cards;
-    if (value == null) return null;
-    if (_cards is EqualUnmodifiableListView) return _cards;
-    // ignore: implicit_dynamic_type
-    return EqualUnmodifiableListView(value);
-  }
-
-  @override
-  final bool isCompleted;
-  @override
-  final CsvMetadata? csv;
-  @override
-  @JsonKey(fromJson: parseError)
-  final String? error;
-  @override
-  final Language? language;
-
-  @override
-  String toString() {
-    return 'GetCardsResponse(sessionId: $sessionId, cards: $cards, isCompleted: $isCompleted, csv: $csv, error: $error, language: $language)';
-  }
-
-  @override
-  bool operator ==(dynamic other) {
-    return identical(this, other) ||
-        (other.runtimeType == runtimeType &&
-            other is _$_GetCardsResponse &&
-            (identical(other.sessionId, sessionId) ||
-                other.sessionId == sessionId) &&
-            const DeepCollectionEquality().equals(other._cards, _cards) &&
-            (identical(other.isCompleted, isCompleted) ||
-                other.isCompleted == isCompleted) &&
-            (identical(other.csv, csv) || other.csv == csv) &&
-            (identical(other.error, error) || other.error == error) &&
-            (identical(other.language, language) ||
-                other.language == language));
-  }
-
-  @JsonKey(ignore: true)
-  @override
-  int get hashCode => Object.hash(
-      runtimeType,
-      sessionId,
-      const DeepCollectionEquality().hash(_cards),
-      isCompleted,
-      csv,
-      error,
-      language);
-
-  @JsonKey(ignore: true)
-  @override
-  @pragma('vm:prefer-inline')
-  _$$_GetCardsResponseCopyWith<_$_GetCardsResponse> get copyWith =>
-      __$$_GetCardsResponseCopyWithImpl<_$_GetCardsResponse>(this, _$identity);
-
-  @override
-  Map<String, dynamic> toJson() {
-    return _$$_GetCardsResponseToJson(
-      this,
-    );
-  }
-}
-
-abstract class _GetCardsResponse implements GetCardsResponse {
-  const factory _GetCardsResponse(
-      {final String? sessionId,
-      final List<AnkiCard>? cards,
-      required final bool isCompleted,
-      final CsvMetadata? csv,
-      @JsonKey(fromJson: parseError) final String? error,
-      final Language? language}) = _$_GetCardsResponse;
-
-  factory _GetCardsResponse.fromJson(Map<String, dynamic> json) =
-      _$_GetCardsResponse.fromJson;
-
-  @override
-  String? get sessionId;
-  @override
-  List<AnkiCard>? get cards;
-  @override
-  bool get isCompleted;
-  @override
-  CsvMetadata? get csv;
-  @override
-  @JsonKey(fromJson: parseError)
-  String? get error;
-  @override
-  Language? get language;
-  @override
-  @JsonKey(ignore: true)
-  _$$_GetCardsResponseCopyWith<_$_GetCardsResponse> get copyWith =>
-      throw _privateConstructorUsedError;
-}
